# GPM/workbench/web/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GPM Workbench API",
    "description": "Pedagogical API over GPM/functions",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/editor/compile": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Compile Endpoint",
        "operationId": "compile_endpoint_api_editor_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor/reconstruct": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Reconstruct Endpoint",
        "operationId": "reconstruct_endpoint_api_editor_reconstruct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconstructRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor/gpm/read": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Gpm Read",
        "operationId": "gpm_read_api_editor_gpm_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpmReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor/gpm/read/file": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Gpm Read File",
        "operationId": "gpm_read_file_api_editor_gpm_read_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_gpm_read_file_api_editor_gpm_read_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor/gpm/write": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Gpm Write",
        "operationId": "gpm_write_api_editor_gpm_write_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpmWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor/gpm/search": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Gpm Search",
        "operationId": "gpm_search_api_editor_gpm_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpmSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor/spectroscope": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Spectroscope Endpoint",
        "operationId": "spectroscope_endpoint_api_editor_spectroscope_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpectroscopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_gpm_read_file_api_editor_gpm_read_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_gpm_read_file_api_editor_gpm_read_file_post"
      },
...
      "CompileRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "nl",
              "code",
              "hybrid"
            ],
            "title": "Mode"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "og"
          },
          "language_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Id"
          },
          "content_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Key"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "CompileRequest"
      },
...
      "CurveMeta": {
        "properties": {
          "full_count": {
            "type": "integer",
            "title": "Full Count"
          },
          "downsampled": {
            "type": "boolean",
            "title": "Downsampled"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 512
          }
        },
        "type": "object",
        "required": [
          "full_count",
          "downsampled"
        ],
        "title": "CurveMeta"
      },
...
      "GpmReadRequest": {
        "properties": {
          "base64": {
            "type": "string",
            "minLength": 4,
            "title": "Base64"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "base64"
        ],
        "title": "GpmReadRequest"
      },
      "GpmSearchRequest": {
        "properties": {
          "document_ref": {
            "type": "string",
            "title": "Document Ref"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "mode": {
            "type": "string",
            "enum": [
              "word",
              "gcd",
              "lcm"
            ],
            "title": "Mode",
            "default": "word"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries"
          }
        },
        "type": "object",
        "required": [
          "document_ref"
        ],
        "title": "GpmSearchRequest"
      },
      "GpmWriteRequest": {
        "properties": {
          "document_ref": {
            "type": "string",
            "title": "Document Ref"
          },
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "og"
          },
          "encrypt": {
            "type": "boolean",
            "title": "Encrypt",
            "default": false
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "document_ref"
        ],
        "title": "GpmWriteRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReconstructRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "nl",
              "code",
              "hybrid"
            ],
            "title": "Mode"
          },
          "document_ref": {
            "type": "string",
            "title": "Document Ref"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "document_ref"
        ],
        "title": "ReconstructRequest"
      },
...
      "SpectroscopeRequest": {
        "properties": {
          "document_ref": {
            "type": "string",
            "title": "Document Ref"
          },
          "token_start": {
            "type": "integer",
            "title": "Token Start",
            "default": 0
          },
          "token_end": {
            "type": "integer",
            "title": "Token End",
            "default": 0
          },
          "content_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Key"
          }
        },
        "type": "object",
        "required": [
          "document_ref"
        ],
        "title": "SpectroscopeRequest"
      },
      "Step": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lines"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "type": "object",
            "title": "Values"
          },
          "formula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formula"
          },
          "extra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "detail"
        ],
        "title": "Step"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkbenchResponse": {
        "properties": {
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Steps"
          },
          "explain_links": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explain Links"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "curve_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurveMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "WorkbenchResponse"
      }
    }
  }
}